refactor: moves logic into hooks module
